requirements.txt
seaborn>=0.12.0
statsmodels>=0.14.0
pandas>=3.0.2
numpy>=2.4.4
holidays>=0.40.0
lightgbm>=4.6.0
pyyaml>=6.0.3
scikit-learn>=1.8.0